1.5.16-beta2
* Tested on Android

+++ PCE.emu/src/mednafen/cdrom/pcecd.cpp
@@ -1139,7 +1139,7 @@ int PCECD_StateAction(StateMem *sm, int load, int data_only)
 	 SFVAR(Fader.CountValue),
 	 SFVAR(Fader.Clocked),
 
-	 SFARRAY(&SubChannelFIFO.data[0], (uint32)SubChannelFIFO.data.size()),
+	 SFARRAY(&SubChannelFIFO.data[0], SubChannelFIFO.data.size()),
 	 SFVAR(SubChannelFIFO.read_pos),
 	 SFVAR(SubChannelFIFO.write_pos),
 	 SFVAR(SubChannelFIFO.in_count),
